Conversión de EspacioMetrico a funciones estáticas. Limpieza y mini-optimización.
Actualización del diagrama de clases UML.

+++ Fuentes/io/ArchivoBloque/Fuentes/ArchivoBloque.cpp
@@ -13,7 +13,7 @@ ArchivoBloque::ArchivoBloque(const char *nombre, size_t tamanioBloque): tamanioB
 	this->archivo = ArchivoFactory_Nuevo(nombre, eTipoArchivo::eTipoArchivo_Binario);
 
 	// nos aseguramos de que el bloque 0 siempre exista
-	if (this->archivo->GetFileSize() < this->tamanioBloque)
+	if ((size_t)this->archivo->GetFileSize() < this->tamanioBloque)
 	{
 		char buff[this->tamanioBloque];
 		for (size_t i = 0; i < this->tamanioBloque; i++)
@@ -50,15 +50,10 @@ iBloquePtr ArchivoBloque::LeerBloque(size_t nroBloque)
 
 void ArchivoBloque::EscribirBloque(size_t nroBloque, iBloquePtr bloque)
 {
-//	if (!bloque->FueModificado())	// si no se modifico, no hay necesidad de guardarlo en el disco
-//		return;
-
 	this->archivo->Seek(nroBloque * this->tamanioBloque);
 
 	this->archivo->Write(bloque->ObtenerContenido(), this->tamanioBloque);
 	this->archivo->Flush();
-	
-//	bloque->BorrarBitModificacion();	// el disco tiene ahora el mismo contenido, por lo tanto, esta actualizado
 }
 
 void ArchivoBloque::Close()
